Make Selenium listener port configurable
The Sauce Connect binary defaults to starting up the Selenium
relay on port 4445, but exposes a flag to override the default
in cases where that port is already in use.  This allows users
to specify a `port` option in `wct.conf.js` to override the
default Selenium listener port when using wct with Sauce
Connect.